requirements.txt
aiosmtplib==2.0.2
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
blinker==1.8.2
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.4
fastapi-mail==1.4.1
h11==0.14.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
motor==3.6.0
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pymongo==4.9.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.16
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0
